pyproject.toml
[project]
name = "uptimer"
version = "0.1.0"
description = "Service uptime monitoring CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "itsdangerous>=2.0.0",
    "pymongo>=4.0.0",
    "croniter>=2.0.0",
]

[project.scripts]
uptimer = "uptimer.main:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pyright>=1.1.350",
    "coverage[toml]>=7.4.0",
    "mongomock>=4.1.0",
    "types-PyYAML>=6.0.0",
    "types-croniter>=2.0.0",
    "respx>=0.21.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require network access)",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
